* change Venue to use Integer, dont need Long as an ID

* change Stage to use Integer, dont need Long as an ID

* add missing venue Integer moves

* add UUID generation via Hibernate

* add UUID to Staff, Person, Patron

* change ID to UUID in Events

* update person ID variable name

+++ backend/src/main/java/com/code/theaterapp/event/dtos/CreateEventDTO.java
@@ -3,7 +3,7 @@ package com.code.theaterapp.event.dtos;
 public record CreateEventDTO(
         String title,
         String showTime,
-        Long stageId,
-        Long venueId
+        Integer stageId,
+        Integer venueId
 ) {
 }
