+++ backend/src/main/java/com/code/theaterapp/stage/dtos/CreateStageDTO.java
@@ -3,6 +3,6 @@ package com.code.theaterapp.stage.dtos;
 public record CreateStageDTO(
         String name,
         Integer capacity,
-        Long venueId
+        Integer venueId
 ) {
 }
